Refs #5 Refactor resource listing for CloudFormation
