+++ package.json
@@ -41,6 +41,7 @@
     "js-yaml": "3.6.1",
     "mocha": "3.0.2",
     "mocha-lcov-reporter": "1.2.0",
+    "rewire": "2.5.2",
     "sinon": "1.17.5"
   }
 }
