+++ src/resource-listers/CloudFormationStackResourceLister.js
@@ -2,90 +2,24 @@
 
 var _ = require('underscore'),
     Q = require('q'),
-    Class = require('class.extend'),
     AWS = require('aws-sdk'),
-    constants = require('../constants'),
-    resourceUtils = require('../util/resource'),
-    cfn = new AWS.CloudFormation(),
-    dynamo = new AWS.DynamoDB();
+    DynamoResourceLister = require('./base/DynamoResourceLister'),
+    cfn = new AWS.CloudFormation();
 
-module.exports = Class.extend({
+module.exports = DynamoResourceLister.extend({
 
    init: function(stackName) {
       this._stackName = stackName;
    },
 
-   listResources: function() {
-      var self = this;
-
+   _fetchTableNames: function() {
       return Q.ninvoke(cfn, 'listStackResources', { StackName: this._stackName })
          .then(function(resp) {
             return _.chain(resp.StackResourceSummaries)
                .where({ ResourceType: 'AWS::DynamoDB::Table' })
                .pluck('PhysicalResourceId')
                .value();
-         })
-         .then(function(tableNames) {
-            return _.reduce(tableNames, function(prev, tableName) {
-               return prev.then(function(resources) {
-                  return self.convertTableNameToResources(tableName)
-                     .then(function(tableResources) {
-                        return resources.concat(tableResources);
-                     });
-               });
-            }, Q.when([]));
-         })
-         .then(function(resources) {
-            return resources;
          });
-   },
-
-   convertTableNameToResources: function(tableName) {
-      return Q.ninvoke(dynamo, 'describeTable', { TableName: tableName })
-         .then(function(resp) {
-            var resources = [];
-
-            resources.push(this.convertTableToResource(resp.Table, constants.READ));
-            resources.push(this.convertTableToResource(resp.Table, constants.WRITE));
-
-            _.each(resp.Table.GlobalSecondaryIndexes, function(index) {
-               resources.push(this.convertIndexToResource(resp.Table.TableName, index, constants.READ));
-               resources.push(this.convertIndexToResource(resp.Table.TableName, index, constants.WRITE));
-            }.bind(this));
-
-            return resources;
-         }.bind(this));
-   },
-
-   convertTableToResource: function(table, capacityType) {
-      return {
-         resourceType: 'table',
-         name: resourceUtils.makeResourceName(table.TableName),
-         tableName: table.TableName,
-         capacityType: capacityType,
-         provisioning: {
-            lastIncrease: table.ProvisionedThroughput.LastIncreaseDateTime,
-            lastDecrease: table.ProvisionedThroughput.LastDecreaseDateTime,
-            numberOfDecreasesToday: table.ProvisionedThroughput.NumberOfDecreasesToday,
-            currentCapacity: table.ProvisionedThroughput[capacityType],
-         },
-      };
-   },
-
-   convertIndexToResource: function(tableName, index, capacityType) {
-      return {
-         resourceType: 'index',
-         name: resourceUtils.makeResourceName(tableName, index.IndexName),
-         tableName: tableName,
-         indexName: index.IndexName,
-         capacityType: capacityType,
-         provisioning: {
-            lastIncrease: index.ProvisionedThroughput.LastIncreaseDateTime,
-            lastDecrease: index.ProvisionedThroughput.LastDecreaseDateTime,
-            numberOfDecreasesToday: index.ProvisionedThroughput.NumberOfDecreasesToday,
-            currentCapacity: index.ProvisionedThroughput[capacityType],
-         },
-      };
    },
 
 });
